Add small and unspaced modifiers to govspeak
These two modifiers are intended to be used in cases where we want to
either have text rendered with govuk-frontend body-small option or if
we want to remove margins from typography. They can also both be used
together.

The use case this was designed for was to allow content designers to put
in govspeak as input for the phase message on services in alpha or beta.
In this situation we want to be able to render a smaller text size and
we don't want to have margins (since the component itself applies
margins). Both of these will only support rather basic govspeak
scenarios and some of the wilder options will likely cause a poor
experience.

I did imagine there may well be other situations when either of these
options may be useful but I would accept that this is rather specific
functionality that may have limited wider use. However modifying this
component felt like an approach that was more consistent with component
principles than having the phase component apply styling.