Update step by step header component style
Removes top border
Changes border bottom colour to blue
Adds margin top of 15px

---

We know that 70% of traffic to coronavirus content is direct to content rather than via landing page. We want to provide users with a clear route to the landing page, and to useful related content. This is to help them discover support that they might not already be aware of or think to search for.

We need to adapt the step by step superbreadcrumb component to be able to be used in this case.

The feature will provide users with a link to the landing page or hub that the content is tagged to using the taxonomy, using the 'superbreadcrumb'.

It will also provide users with a set of related links which will be drawn from the curated content associated with the lowest level topic that the content item is tagged to.

Providing users with clear routes to discover useful content relevant to them.

https://trello.com/c/IOjnO55X/196-update-superbreadcrumb-component-ready-for-coronavirus-tagged-content