Undo recursive FAQPage presenter
I'm intending to do this more simply and specifically for HTML publications
within government-frontend. This code is quite hard to debug without context,
and it's not working quite right at the moment.

Note that this is not a breaking change - it still resolves a valid schema
on HTML publications, but they are less useful.

This reverts commit f0e470a27660e1d0bcfbd318e4957ba983191095.