Add explicit config for Dependabot
Currently Dependabot is enabled for all Gems and packages on this repo.
We'd like to trial limiting the number of PRs opened by Dependabot to
libraries that we care about explicitly.

See https://github.com/alphagov/content-publisher/pull/2062 for more
context on this configuration file (and a similar config file for
Content Publisher).

There's a few reasons I've picked this repo to try out this config (as
well as the existing configuration in Content Publisher).

- It's a Gem, allowing us to learn more about if there's anything
special about Gems which means going down this route would be
problematic.
- It includes a framework library (`govuk-frontend`) which we'd want to
have enabled for this repo, but perhaps not for any others.
- It allows us to test the idea of having a global config file
containing rules for libraries not used by this repo, and check that
Dependabot is okay with that.