Add print link component
This component adds a visual link that includes a print icon and can be
used to help users print a page. This pattern was orginally implemented in
the Government Frontend application, however we'd like to re-use this
component in Smart Answers.

The component adds either a button or link depending on whether the
application support printing via navigation (used to add support for
environments where JavaScript is disabled e.g. /print). The default for
this component is to use a button and rely on JavaScript to call the DOM
API. This allows the component to be added to an application without having
to add extra routes to trigger printing. However for applications that
have routes for printing (e.g. Government Frontend), a `href` can be
specified and the component will render as anchor tag.