Add print link component
This component adds a link that includes a print icon and can be used to
help users print a page. This pattern was orginally implemented in the
Government Frontend application, however we'd like to re-use this
component in Smart Answers.

This implementation of the component relies on JavaScript by calling the
browser print action via js in the href. We considered using a click event
handler, however wasn't sure what then to make the href - as some
browsers have different behaviour with links without hrefs. To
support applications (e.g. Government Frontend) with print links that do
not need JS i.e. `href=blah/print`, this component allows a custom href
to be defined and makes sure the link is displayed in non-js
environments.