Turn on max-nesting-depth stylelint rule
This rule aims to prevent SCSS that is heavily nested by restricting
nesting to a maximum depth of 2. This is probably easiest explained with
an example:

```
.root-rule {
  .first-nesting {
    .second-nesting {
      .third-nesting {
        content: "violation";
      }
      @rule with-declarations {
        content: "violation";
      }
    }

    @rule without-declarations {
      .second-nesting {
        content: "not-violation";
      }
    }
  }
}
```

This could prove to be a somewhat divisive rule as it can be tricky to
work out what to do when you're caught at the limits of it. However if
you don't rely on SCSS nesting (as, I understand, is our preference)
it's pretty easy to work with. For example, if we wrote the above out as
actual rules it shows some smells in the selector choices

```
.root-rule .first-nesting .second-nesting .third-nesting {
  content: "not-violation";
}

.root-rule .first-nesting .second-nesting {
  @rule with-declarations {
    content: "not-violation";
  }
}

.root-rule .first-nesting {
  @rule with-declarations {
    .second-nesting {
      content: "not-violation";
    }
  }
}
```

Applying this rule was rather difficult for govspeak code as that relies
on nesting. I decided it was simpler to turn off the linting rule there
for now, although in theory it should be solvable there as we shouldn't
really need any nesting in there that exceeds the max-depth of 2.