Make public layout template adjustments
Adjust public template layout to be more in line with design system
guidance.
This change separates govuk-width-container and govuk-main-wrapper,
which used to be applied to the same element previously.
This is because elements like the phase banner, back link and
breadcrumbs should sit inside the width container, but outside of the
main page wrapper, in order to achieve the correct layout.