Remove unnecessary dependencies
This does a review of the gemspec file and moves and removes things
accordingly.

This removes rake from the production dependencies as it didn't seem needed
outside the development environment.

It also removes gds-api-adapters from production dependencies as it
seemed to only be used in development. I also didn't see the value in
maintaining the ">= 0" pin for gds-api-adapters.

Foreman doesn't seem to be referenced at all in the source code or
documentation, I assume this is something that was once used with the
development VM. This is good to remove as Foreman recommends not
installing the gem via bundler [1].

The selenium_webdrivers gem is removed as it is a dependency of
govuk_test and isn't used explicitly.

[1]: https://github.com/ddollar/foreman#installation