Simplify Rakefile
The Rakefile is only expected to be run in environments where
development dependencies are available. Because of this we can drop the
rescue for loading the rspec gem.

This then switches the Rubocop approach from being a shell out to
instead using the rake task that rubocop provides.

Finally it consolidates the distinct lint tasks into a single one that
does all linting (Ruby, SCSS and JS).