Make gem loading more predictable to fix PhantomJS quirks
We thought we had replaced PhantomJS fully in
https://github.com/alphagov/govuk_publishing_components/pull/1598
however it was only removed in certain circumstances and it is still
executing when running the default `rake` task.

When I looked into the issue I discovered that the
`Bundler.require(*Rails.groups)` loaded different gems depending on the
task used (running the default rake task for instance did not load all
the expected gems).

I decided that the simplest approach would be to remove this line and
load all the gems manually. This felt like a less surprising approach
for a rubygem and less surprising for users in why gems were loaded.

In order to have Chrome selenium load correctly we need to ensure that
the govuk_test gem is loaded first (to introduce selenium-webdrivers),
then we can include jasmine_selenium_runner.

I also removed the rescue block around Jasmine as it isn't expected for
this file to be included without development gems installed.