Fix JS test that fails with chromedriver
Chromedriver via Selenium seems to have a problem with tests that rely
on focus or blur states being triggered automatically. In order to test
them we need to synthesise these events.

We experienced similar problems with Content Publisher [1] and resolved
them with synthetic events as I have done here. From searching I can't
find a definitive reason for this (or much information at all), this
puppeteer issue [2] seems to be the closest. My personal theory is that
focus involves integration with the OS UI which doesn't occur when
running in a headless mode.

[1]: https://github.com/alphagov/content-publisher/issues/702
[2]: https://github.com/puppeteer/puppeteer/issues/1462