Don't load JS related test gems in Heroku
When this app on Heroku chrome and chromedriver are not available
without additional buildpacks. Since we don't have usage for chrome in
Heroku it's simpler to use not load these gems in.

It's a somewhat strange situation how this app runs on Heroku, as we
need the app to be in development mode (so that we've got the various
gems needed to run the dummy application) but we're not going to
actually use all the gems when running the app. This seems the simplest
approach to take for now as opposed to setting some complex gem
groupings.