Add govspeak to radio component description
Wrapped the radio component's description in a govspeak component so that any
markup used in the description is properly displayed. An example in the docs
has been added for this use case, and the test has been updated.

The radio component wrapped the description in a `div` element that has a
`govuk-body` class. This is okay if there's only a single paragraph of text
for the description; but the correct styles aren't added if markup is used in
the description.

This happens on several of the simple smart answers in `frontend`.

Wrapping the description in a govspeak component and a `govuk-body` class
means that both strings that doesn't contain any markup and strings that do
contain markup will inherit the appropriate styles. Thanks CSS!

Changing the way the component accepts a description wasn't an option as this
flaw is already being used in production - the same `description` parameter
needs to be used for markup _and_ for text.

Regex could've been used to test whether the string started with an HTML
element (for example: `/^<[a-z]>/`), and then only use the govspeak component
if there was markup present - but I thought that was too conservative in what
the description parameter could successfully accept.