Specify lang for image card component
Introduce the ability to pass an (optional) lang parameter to specify
language on the image card component.

This is necessary due to the fact that there are instances on the site
where a page will be in Welsh, but certain components on the page will
display data in English as there is no Welsh translation for those parts
of the page. This is a Web Accessibility fail.