Feedback component lang
As of right now, the feedback component is not localised.

On some translated pages, most of the content will be in the specified
language, however the feedback component is in English and not clearly
marked as such. This is a WCAG fail.

This adds lang="en" to the feedback component to correctly indicate the
language the component is in. This is subject to change if/when we end
up translating this component into other languages.