Rework panel component disjointed heading
The prepended text, heading text, and appended text on the panel
component should be in the same block.

The issue with the prepended and appended text being outside the heading
element is that they generally provide context for the heading. Screen
reader users might be confused by the incomplete heading and not
understand it out of context when they are navigating a page according
to its headings.