Search component black background variant
A search revealed that the only place the "no_border" variant of the
search component was being used is as part of the layout header over a
black background.

The search component requires different styling when it is over a dark
background (with further differences between black and blue variations).

This removes the "no border" variation, replacing it with a "black
background" variation, and refactors some of the CSS.