Feedback form errors frontend workaround
When invalid data is submitted, the server responds with a 422 status.
However instead of the intended JSON response, it returns a rendered
html page. The frontend defaults to the generic error message which is
very vague and fails WCAG.

This is tricky to fix from the backend side right now as there is
clobbering by nginx on all 422 requests. This implements a frontend
workaround for the time being. If a JSON response is not present, and
the status code is 422, the frontend generates the error messages that
would have otherwise been sent over from the backend.

Also see https://github.com/alphagov/feedback/pull/758 for the original
work to address this issue