Add default aria-label for contents list component
Updates aria label for the contents list component so that aria label
defaults to "Contents" unless a custom aria label is supplied.

The contents list component is a nav element, and appears on pages which
have other nav elements. When there are multiple navigation landmarks on
a page, they should be labelled to help assistive technology users
differentiate between them