Add aria-label to government_navigation component
Add a relevant aria-label to the government nav component.
Uses the suggested name of "Departments and policy navigation". This was
suggested because in the site footer the section containing the same
links is called "Departments and policy"

The reason for this change is that the site header links are a nav
element and often appears on pages where other nav elements are present.
When there are multiple navigation landmarks on a page, they should be
labelled in order to help assistive tech users differentiate between
them.