Extend contents list component
Introduce the ability to pass through lang and custom title to the
Contents list component.

If a title lang is not supplied, a lang attribute will not be added.
Likewise, if a custom title is not specified, the component will default
to using "Contents" as a title.

The need for this arose because in some instances (for example on world
location pages), when viewing a translation of the page, this title of
this component will be in English and not marked as such.
If a translation exists, we should use the translated string as the
title (e.g Contenu if the page is in French). Otherwise we can use the
English fallback, as long as lang="en" is specified.