Adjust subscription links component styling
There are instances where the "subscribe to feed" link functions as a
button.
This discrepancy between the appearance and the function of this element
can make it more difficult for screen reader and voice control users to
find the information they need.
This changes the styling of the subscription links component when the
"subscribe to feed" element functions as a button. Adjust the markup of
the element to have the role="button" attribute. Also adjust the styling
for the "Get email alerts" link for when the two are side by side, to
link the two together visually.