Force contents list  title to always be Contents or regional equivalent
For consistency we are moving to always having the contents list title
be "Contents". There are some examples in Whitehall where a translated word
for Contents is passed in, along with a language indicator. This moves those
translations into the component and handles the language detection, removes
the ability to hide the title, and removes redundant documentation.