Fix orange outline on global site search
The global header was recently updated to use the search component
instead of a hard coded search field.

This resulted in an issue where the input outline is now orange due to
more specific styles from govuk_template conflicting with the search
component styles.

This updates the component styles to be more specific in order to
override the CSS from govuk_template. It uses the compatibility mixin to
ensure the fix is only applied when in compatibility mode, and also to
ensure the code is easier to clean up when compatibility mode is no
longer a "thing".