Fix chart glitch when source table has row headers
Govspeak can now generated tables with row headings using the correct `th`
element. Previously these were unavailable, which led to a lot of tables that
couldn't be understood by assistive technologies such as screenreaders.

But using a `th` as a row heading broke Magna Charta's automatically generated
chart - the `th` would be ignored and the first `td` used instead. This has led
to broken charts.

This fixes that problem by making sure that both `th` and `td` are use when
looking for table cells when generating the chart.

It adds an extra test to make sure that row headings can be used; and keeps the
old test to make sure that nothing breaks.

The documentation has been expanded to include the multiple headings table.