Move initial-focus script to lib
The initial-focus script is being used by both success alert and error alert component, but because it doesn't use the same name with the components it's not being picked up by the suggested imports, hence missing from all apps using the component. Moving it to lib will make it available to every application.