Add heading level to attachment component
Introduce ability to override default heading_level on the attachment
component.

The component is currently hard coded to use h2 which isn't very
flexible. This component is often used as a child of a section of a
page, in which case the heading level of the component should be
adjusted to reflect the hierarchy of the content.