Switch to standardx
Standard 16 introduced a rule that disallowed the use of `var`, instead
preferring `let` or `const` [1]. This conflicts with the GOV.UK approach
where we tend to not embrace features that we know will break old
browsers even if they're not necessarily supported [2], disallowing var
will mean that < IE 11 will be unable to run any of the JS.

In order to customise standard rules this project has switched to using
standardx [3] which allows us to disallow rules. I've used this so we
can disallow the 'no-var' rule.

[1]: https://github.com/standard/standard/issues/633
[2]: https://github.com/alphagov/govuk_publishing_components/pull/1611#discussion_r456493365
[3]: https://github.com/standard/standardx