Fix contextual sidebar's Brexit countdown lang bugs
The countdown has the page's language whether or not there was a translation
available - there is now a check in place to make sure that there is a
translation available to see whether a `lang` attribute is needed.

The countdown was inheriting a page's right-to-left direction, which would make
the countdown display the numbers in the wrong order - so 39 became 93. The
`dir` attribute that controls the text direction has been set to `ltr` to avoid
this problem.

The translation helpers were previously being used by the related navigation
component - they've now been moved to the shared helpers and tests have been
added to ensure that they work as expected.