Replaces the random placekitten image placeholder
Whilst place kitten is the second best placeholder service available, it
supplies random images - which doesn't play well with visual regression tests.
Replacing it with a non-random image means that the visual regression tests
won't flag a page as having changed.

The example image used for the large image was 404-ing, so I replaced it.