Amend share links columns spacing
When increasing the font size to 200%, some of the social media links
are overlapping, making them unreadable for users who use this setting.

This changes the column widths to use em values instead of px which
means that the width of the link containers increase proportionately to
the font size, which should make for a better experience for users who
use the text zoom setting in their browser.