Update subscription links default text
This has been an agreed content change for emails links as we move away
from the terminology 'alert'.

Also removes the `default` argument as I'm struggling to see why we need
it in this instance. Using `default` also meant that a misconfigured locale
path was being covered up which has been corrected.

Trello:
https://trello.com/c/ssSnzp2L/640-remove-alert-from-all-the-sign-up-links