Add type="button" to Show/Hide password button
This is aiming to fix a bug where in the context of a form that contains
a show password component, when the user hits Enter on any of the input
fields in the form, the browser toggles the passwords visibility
(instead of the expected behaviour, which would be to submit the form).

The browser assumes that if you press the enter key in any of the
fields, you want to submit the form and fires the click event on the
first button, even when the button is not of type="submit". On a form
that contains this component, the 1st button could be the Show/Hide form
button, hence this bug.

Adding type="button" attribute to the show/hide element would resolve this
issue, as it would now be clear to the browser that this is not a submit
button.