Replace countdown component with sidebar partial
Previously we had a separate component to represent the countdown
functionality, which was only used as part of the sidebar. Once
the countdown has elapsed, the bulk of the component will become
redundant, so it makes sense to incorporate it directly into the
contextual sidebar since:

- It's not intended for re-use - the design is coupled to how it
appears as part of contextual navigation.

- It's no longer a countdown, and no other naming is appropriate,
given its single use case.

Note this preserves all of the other behaviour of the component,
including the button/focus.