Layout header nav custom aria label
Introduce the ability to pass a custom aria-label for layout header
navigation. Set the default aria label to "Top level" – this was
previously set on the nav element's child ul, however it makes more
sense to add the aria-label to the nav itself, for screen reader users
who navigate the page by navigation landmarks.