Update data-module on layout header and footer
Change value of data-module attribute to "gem-track-click" on the layout
header and layout footer components.
The reason behind this change is that click tracking is not working on
these elements on GOVUK Accounts. This is because Accounts does not use
static and therefore also does not use the track-click script which
enables click tracking on modules with the data-module="track-click"
attribute.

Changing the value of this attribute to "gem-track-click" means that the
header and footer will work with the gem's track-click script instead.