Add custom heading level option to radio component
Currently if the radio component is passed a heading, and the heading is
not a "Page heading", then that heading is always a h2.
This is not always appropriate; there are instances where more
flexibility is necessary in order to maintain a correct hierarchical
heading structure.
This introduces the shared helper's heading level function to the radio
component. This way we can ensure the default remains h2, while also
having the ability to pass a custom heading level where necessary.