Update header search button to be more specific
Having the heading search button labelled "Search" clashed with a lot of the
finders, which also had a button labelled "Search" on it. This is beause screen
readers can collate and read out form fields, so having duplicate submit button
text could lead to confusion. Changing this to "Search GOV.UK" makes the intent
of this form clearer.