Add Details module from GOVUK Frontend
As part of the Accounts team DAC report it was flagged that the details
element was not functional in Internet Explorer and assistive tech was
not able to interact with this component as expected.

Some investigation revealed that:
1. this is well known issue where all versions of Internet Explorer and
   Edge (pre-chromium) don't fully support details/summary:
https://caniuse.com/details

2. on the Design System's version of the details component, the issue
   would not occur, which would imply that our components gem version is
not properly importing the fix from GOVUK Frontend.

After some digging I realised that while we are requiring the
govuk-frontend's details.js script in our own component javascript, it
is never initialised, therefore we were not applying their polyfill for
IE.
These changes add GOVUK Frontend's Details JS module to our own GOVUK
Modules object so that the module can be initialised with the other
modules, which should automatically resolve our IE issue.