Amend JS modules initialiser code
The way the module start function works currently is that it will either
init GOV.UK Publishing & Legacy Modules, or it will init the GOV.UK
Frontend equivalent.
This does not work well in situations where we need both, for instance
in the case of the details component, where we have a GovukDetails.js
and a Details.js (from GOV.UK Frontend) which both need to be run for
the details component to have full functionality.
What is happening currently, is that:
- the details component has data-module="govuk-details", therefore in
the modules.js start function, moduleName is set as GovukDetails and
frontendModuleName is Details
- the first condition in the if statement evaluates to true because
GOVUK.Modules[GovukDetails] exists and it has an init function, and it
has not been started yet
- since the condition before it evaluated to true, the `else if` below
  does not run. As a result, the GOV.UK Frontend Details module is never
initialised

This breaks up the if-else-if statement into two separate if statements,
which will ensure that both can run if the conditions evaluate to true.