Rename i18n usage from 'govuk_component' to 'components'
It's not clear how this was working before - we suspect govuk_component
is a legacy term - and that it has always called the underlying
components key regardless. The i18n-coverage task was complaining about
unused keys, which included all of these, which naturally looked as though
they were not being used because they were missing the 'components.' prefix.

Tests continue to pass and locally checking out the component gallery suggests
it's all continuing to work as it should.