Add legacy `govspeak` class alongside gem govspeak
There are many many many places where the older `govspeak` class is used on
elements that wrap around content, rather than `gem-c-govspeak`. This class
comes as part of the content from the content store (eg [in `details.body` in
the cabinet office content][1].

Adding the legacy class as a fix is easier and safer that changing the myriad of
places where this is added - around 7K places across alphagov[2]).

This makes sure that the `govspeak` has the same output as `gem-c-govspeak` for
minimal additional CSS.

[1]: https://www.gov.uk/api/content/government/organisations/cabinet-office
[2]: https://github.com/search?q=org%3Aalphagov+%27%3Cdiv+class%3D%22govspeak%22%3E%27&type=Code&ref=advsearch&l=&l=