Retire is_page_heading parameter (part 1)
This is step 1 towards removing the is_page_heading parameter from the
radio component. The reason behind this is that the ability to pass a
custom heading level was introduced to the radio component some time
back #1951
The is_page_heading parameter is a specialised parameter that dictates
that the heading should be a h1 of size xl.
This used to solve a problem when the component did not use to support
heading_level, however now that we can use heading_level and
heading_size to achieve the same result, the is_page_heading parameter
becomes redundant.

The reason we have to do this in steps is that the component is
referenced with the is_page_heading parameter in several applications.
Due to the way the logic in this component is structured, it would be
a breaking change if we simply went ahead with replacing
is_page_heading: true with heading_level: 1 and heading_size: 'xl'