Remove is_page_heading param from radio component
Recently, the ability to pass a custom heading level to the radio
component was introduced.
Since a custom heading level is now an option, the is_page_heading
parameter became redundant. This parameter simply dictates that the
radio component legend heading should be a h1 of size xl by default.
It is possible to achieve the same result using heading_level and/or
heading_size instead.

This removes all mentions of is_page_heading from the radio component.
Whenever we need this component to have a h1 in the future we can simply
use the heading_level option (in combination with heading_size, if
necessary)