Fix IE11 initCustomEvent error
Some work was done to allow cookie consent dependent modules to start
immediately when the user consents to cookies.
To make this possible, some functionality was introduced to trigger a
custom cookie consent event when consent is given. It seems like some of
the code that was introduced causes a JS error on IE11 specifically. The
error appears to be related to the initCustomEvent method. Namely, it
seems like IE11 does not like when this method is not called with less
than 4 arguments, and consequently throws an error.
This passes null as a 4th argument, which should appease IE11 without
causing issues in other browsers.