Review inverse heading guidance and examples
A few examples here use old markup – so I have updated them with what's currently generated by the breadcrumbs and title components to check they work well together. We're also adding an example on how to the inverse modifier in case we need to pass a link in the block.