Added tracking for the header menu toggle
There is already tracking in place to see how users use the search
toggle - this adds tracking for the menu toggle.

Since the functionality for the toggling of the menu comes from GOV.UK
Frontend, this needed to be a separate event listener added to
specifically send events to analytics.

This can be removed when the user research has been completed.