Update action link dark arrow
Currently the dark version of our action link icon is using yellow as it was initially only used on C19 pages. Now with extending the usage of this component on the Brexit landing page we shift it to a neutral white so that it can be used on both campaigns.