Use JavaScript module prefixes consistently
For components that have scripts coming from `govuk-frontend` and `govuk_publishing_components` we used to have a Govuk prefix attached to the module name (e.g. GovukCheckboxes). This was also used as a workaround in the modules system to initialise both scripts – one using the GOV.UK Publishing pattern, the other using GOV.UK Frontend pattern. However, when moving to a single module system (th GOV.UK Frontend pattern) this workaround couldn't be used anymore. This forces us to reconsider how multiple modules get attach to a single element. The first step is to avoid naming confusion, so we now use, for example, GovukCheckboxes and data-module='govuk-checkboxes' for scripts coming from `govuk-frontend` (as the `govuk-` prefix is consistently used across SCSS mixins styles and markup) and GemCheckboxes and data-module='gem-checkboxes') for scripts defined in `govuk_publishing_components`.