Allow multiple modules on a certain element
The second step is to allow multiple modules to be applied to a single element. The best example for illustrating the need for this is the details script. Other components sharing scripts from both libraries, such as the checkboxes component or a combination of UI enhancements and tracking could also use this feature.