Add faux `require_tree` for Govspeak and polyfills
A recent change meant that `include` was needed for individual
JavaScript files in the polyfill[1] and govspeak[2] folders - this was
to avoid Static _and_ frontend applications from loading the same
JavaScript file twice on the same page.

But this means that any additions to these folders will be left out and
have to manually added to each of the frontend applications again.
Unfortunately `require_tree` is only available for local files - so the
frontend applications can not `require_tree` for folders in the gem.

This pull request adds a `all.js` file in each folder that uses
`require_tree` for the folder - so it can be used instead of
`require_tree` in each frontend application, doesn't need manually
updating, and ensures that any additions to each folder is not
forgotten.

[1]: https://github.com/alphagov/frontend/blob/b9c48fb1dc04901ba7f8ae92076d2dfeb1c4240e/app/assets/javascripts/application.js#L4-L6
[2]: https://github.com/alphagov/frontend/blob/b9c48fb1dc04901ba7f8ae92076d2dfeb1c4240e/app/assets/javascripts/application.js#L21-L23