Switch environment app helper to use GOVUK_ENVIRONMENT_NAME
This commit switches the `Environment` app helper to use the `GOVUK_ENVIRONMENT_NAME` when getting the current environment for use in the layout header component. The reason for this change is that we're removing the `ERRBIT_ENVIRONMENT_NAME` environment variable in favour of `GOVUK_ENVIRONMENT_NAME`, with the former being an overhang from the days when GOV.UK used Errbit. The `GOVUK_ENVIRONMENTS` variable is refactored from a hash to an array, as we no longer have a weird edge case where the name of the integration environment is different to the acceptance environment that the method returns; there will be a performance impact with this as we're no longer doing a constant-time lookup, but given the number of environments this should be negligible.

It also refactors the associated tests and introduces coverage for the Heroku environment.