Added dropdown menu content
Changed the shape of the dropdown menu contents in the YAML snippet to
be one array per drop down menu; originally it was split into multiple
arrays as the visual layout demanded columns of links. However, this
meant that the list was truncated when read aloud - for example "list,
7 items" then "list, 7 items" rather than "list, 14 items". This also
means that the HTML makes more sense with a list inside a list item.

Since the showing and the hiding of the drop down menus will use
JavaScript they are hidden by default.

The fixture in the test has been updated with the dropdown menus.

Add search dropdown menu & popular links.