Update custom dimension
The intention of CD111 is to track page views for pages displaying
the brexit superbreadcrumb. The current code is tracking any page
tagged to a brexit priority taxon, irrespective of if the breadcrumb
is displayed. Eg. the /import-goods-into-uk step by step.

Happily, this data is also required, but it should be assiged to 112.
Amending the code to track page views for pages displaying
the brexit superbreadcrumb in the next commit.