Give working examples for the checkbox name
Checkboxes in Rails have a bit of a configuration quirk. If you don't
add `[]` to the name, only the last selected value is passed to the
controller, rather than an array of all of the selected values.

The name needs to be backed by indicating that the checkbox param is an
array in the controller.

For example, if the checkbox is called `nationality[]`, then you need to
do the permit as:

```
params.permit(nationality: [])
```