Update notice component design to match DS
The GOVUK version of the notification component is different from the
Design System version.
Adopt classes and styles from the Design System notification banner
component, removing redundant CSS. Introduce a flag to display an
additional title above the banner, which will default to "Important" but
can be overruled depending on the situation.

This is a step towards the goal of bringing the notice, success, and
error components more in line with each other (and with the DS
equivalents).