First draft of dropdown menu and top level menu
This looks janky as the styles are not complete - but the mechanics
work.

Menu now toggles between:
 - Top level navigation menu and search menu on small screens
 - Second level navigation menu and search menu of larger screens

Second level menu on smaller screens acts as an accordion - multiple
sections can be open at once. One larger screens the second level menu
is a dropdown, so only one section at a time can be open. When the
window is resized from small to large, the last opened section is kept
open whilst the other sections are closed.