Delete empty print stylesheets
There was some refactoring to remove display: none rules from component
stylesheets in favour of using the govuk-!-display-none-print class
instead. alphagov/govuk_publishing_components#1561

As a consequence of that, some of the component print stylesheets which
previously used to contain CSS are now empty files.

Step 1 in the process of deleting these empty files was to hunt down
references of these files in application print stylesheets and remove
them.
Step 2 is removing these files from the components gem altogether.