Fix cookie banner issue (IE10)
It seems that IE10 does not support the hidden attribute. This is
causing some issues with the cookie banner script: when the confirmation
banner is visible (upon rejecting/accepting cookies), the cookie choice
panel should disappear, which does not happen in IE10.
This fixes the issue by using CSS to hide elements accordingly when the
hidden attribute is present (as opposed to toggling display with JS).
Additionally, switch from using property dot notation to set hidden to
true/false, to using setAttribute because even though the former worked
in all major modern browsers, IE10 seemed to ignore it. This is perhaps
due to the distinction between properties and attributes.